arm/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
ccxt
python-dotenv
requests